Fix : MPEG, FLV Splitter - improve H.264 parse for the correct calculation of the video size, ticket #9. Fix : FLV Splitter - H.264, correct calculate Aspect Ratio. Fix : Some C4018 warning;
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@3674 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/apps/mplayerc/FGManager.cpp
@@ -869,7 +869,7 @@ HRESULT CFGManager::Connect(IPin* pPinOut, IPin* pPinIn, bool bContinueRender)
 			psde->mts.AddTail(CMediaType(*pmt));
 		}
 		EndEnumMediaTypes(pmt)
-		if (skip < psde->mts.GetCount()) {
+		if (skip < (int)psde->mts.GetCount()) {
 			m_deadends.Add(psde);
 		}
 	}
@@ -1312,7 +1312,7 @@ STDMETHODIMP CFGManager::GetDeadEnd(int iIndex, CAtlList<CStringW>& path, CAtlLi
 {
 	CAutoLock cAutoLock(this);
 
-	if (iIndex < 0 || iIndex >= m_deadends.GetCount()) {
+	if (iIndex < 0 || iIndex >= (int)m_deadends.GetCount()) {
 		return E_FAIL;
 	}
